requirements.txt
fastapi
uvicorn
openai
pdfplumber
python-docx
python-multipart
